CatsDogsTest

+++ src/main/java/com/ward_n6/service/DogService.java
@@ -1,19 +1,21 @@
 package com.ward_n6.service;
 
-import com.ward_n6.entity.pets.Cat_2;
 import com.ward_n6.entity.pets.Dog_2;
 import com.ward_n6.enums.PetsSex;
-import com.ward_n6.repository.CatsCrud;
 import com.ward_n6.repository.DogsCrud;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import javax.annotation.Resource;
+import java.util.List;
 import java.util.Optional;
 @Service
 public class DogService {
     @Resource
     private DogsCrud dogsCrud;
+
+    public DogService(DogsCrud dogsCrud) {
+        this.dogsCrud = dogsCrud;
+    }
 
 
     public void addDog(PetsSex petsSex, Dog_2 dog_2) {
@@ -26,23 +28,15 @@ public class DogService {
         return byId.orElse(null);
     }
 
-    public String deleteDog(long id) {
-        try {
-            Dog_2 dog = findDog(id);
-            dogsCrud.deleteById(id);
-            return dog.getPetName() + " удалена";
-        } catch (EmptyResultDataAccessException e) {
-            return "Такая собака была удалена ранее или вы ошиблись с поиском.";
-        }
+    public Dog_2 deleteDog(long id) {
+        Dog_2 dog = findDog(id);
+        if (dog != null) dogsCrud.deleteById(id);
+        return dog;
     }
 
 
-    public String allDogs() {
-        StringBuilder stringBuilder = new StringBuilder();
-        for (Dog_2 dog2 : dogsCrud.findAll()) {
-            stringBuilder.append("\n").append(dog2);
-        }
-        return stringBuilder.toString();
+    public List<Dog_2> allDogs() {
+        return dogsCrud.findAll();
     }
 
     public Dog_2 change(long id, Dog_2 dog_2, PetsSex petsSex) {
